More efficient workflow, add threading

+++ services/archive.py
@@ -1,4 +1,5 @@
 import json
+import time
 from services.abstract_service import AbstractService
 
 class Service(AbstractService):
@@ -10,22 +11,29 @@ This keeps the crew chat clean
         super().__init__(*args, **kwargs)
         self.processed: list[str]
         self.archive_path = f"services/{self.crew.name}-archive.json"
-
-    def on_post(self, posts):
         try:
             with open(self.archive_path, 'r') as file:
-                archive = json.load(file)
+                self.archive = json.load(file)
         except:
-            archive = []
-        for post in posts[10:]:
-            if post["type"] == "wrote_crew_wall_message" and post["content"].startswith("/"):
-                for comment in post["comments"]:
-                    if int(comment["actor"]["id"]) == self.connector.get_self_player().id:
-                        self.connector.delete_post(post_id=post["id"])
-                        archive.append(post)
-                        break
-        with open(self.archive_path, "w") as file:
-            json.dump(archive, file)
+            self.archive = []
+        
+        self.processed_since_last_save = 0
+
+    def on_post(self, post):
+        
+        if post["type"] == "wrote_crew_wall_message"\
+        and post["content"].startswith("/")\
+        and post["time"] + 600 < time.time():
+            for comment in post["comments"]:
+                if int(comment["actor"]["id"]) == self.connector.get_self_player().id:
+                    self.connector.delete_post(post_id=post["id"])
+                    self.archive.append(post)
+                    self.processed_since_last_save += 1
+                    break
+        if self.processed_since_last_save > 10:
+            self.processed_since_last_save = 0
+            with open(self.archive_path, "w") as file:
+                json.dump(self.archive, file)
 
     def start(self):
         self.running = True
